+++ src/entities/apocalypse/index.ts
@@ -1,1 +1,0 @@
-export { apocalypses, apocalypsesCategories } from './model/apocalypses';