+++ src/entities/apocalypse/model/apocalypses-kits/apocalypses-borov-edition-kit.ts
@@ -1,4 +1,0 @@
-import { Apocalypse } from '@/src/shared/lib/types/apocalypse';
-
-// const previousKitLastId: number = apocalypsesStandartEditionKit.at(-1)?.id || 0;
-export const apocalypsesBorovEditionKit: Apocalypse[] = [];