+++ src/entities/profession/index.ts
@@ -1,1 +1,0 @@
-export { professions } from './model/professions';