+++ src/entities/profession/model/professions-kits/professions-standart-edition.ts
@@ -1,3 +1,0 @@
-import { Profession } from '@/src/shared/lib/types/profession';
-
-export const professionsStandartEdition: Profession[] = [];