+++ src/entities/shelter/index.ts
@@ -1,1 +1,0 @@
-export { sheltersCategories } from './model/shelters';