+++ src/entities/shelter/model/shelter-name-to-shelter.ts
@@ -1,10 +1,10 @@
 import { Shelter } from '@/src/shared/lib/types/shelter';
-import { shelters } from './shelters';
+import {  sheltersStore } from './shelters';
 
 export function shelterNameToShelter(shelterName: string): Shelter {
-  const shelter = shelters.find(shelter => shelter.name === shelterName);
+  const shelter = sheltersStore.getShelterByName(shelterName);
 
-  if (shelter === undefined) {
+  if (!shelter) {
     throw new Error('Shelter is undefined');
   }