+++ src/layouts/app-layout/ui/ui.tsx
@@ -5,15 +5,12 @@ import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { SafeAreaProvider, SafeAreaView, useSafeAreaInsets } from 'react-native-safe-area-context';
 import Constants from 'expo-constants';
 import { ImageBackground } from 'expo-image';
+import { databaseStore } from '@/src/shared/model/database-store';
+import { useSQLiteContext } from 'expo-sqlite/next';
 
 type PropsType = {
   children: React.ReactNode;
 };
-
-const SCREEN_HEIGHT = Dimensions.get('screen').height; // device height
-const STATUS_BAR_HEIGHT = StatusBar.currentHeight || 24;
-const WINDOW_HEIGHT = Dimensions.get('window').height;
-const NAVIGATION_BAR_HEIGHT = SCREEN_HEIGHT - WINDOW_HEIGHT - STATUS_BAR_HEIGHT;
 
 export function AppLayout({ children }: PropsType) {
   const [fontsLoaded, fontsError] = useFonts({
@@ -25,7 +22,10 @@ export function AppLayout({ children }: PropsType) {
     SpaceMono: require('@/assets/fonts/SpaceMono-Regular.ttf'),
   });
 
+
+  databaseStore.database = useSQLiteContext();
   useEffect(() => {
+
     setTimeout(() => {
       StatusBar.setTranslucent(true); //TODO find component that overriding the properties
       StatusBar.setBackgroundColor('#c3b5a8');
@@ -35,6 +35,7 @@ export function AppLayout({ children }: PropsType) {
   if (!fontsLoaded && !fontsError) {
     return null;
   }
+
 
   return (
     <SafeAreaView style={{ flex: 1 }}>
@@ -54,8 +55,8 @@ export function AppLayout({ children }: PropsType) {
 
 const s = StyleSheet.create({
   background: {
-    height: '100%',
-    width: '100%',
+    height: "100%",
+    width: "100%",
   },
   root: {
     height: '100%',