+++ src/pages/create-game-page/ui/ui.tsx
@@ -1,6 +1,6 @@
 import { ImageBackground, SafeAreaView, ScrollView, Text, View } from 'react-native';
 import { Footer } from '@/src/shared/ui/footer/ui';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Header } from './header/ui';
 import { GameOptionCheckbox } from '@/src/pages/create-game-page/ui/game-option-checkbox/ui';
 import { GameOptionSelect } from '@/src/pages/create-game-page/ui/game-option-select/ui';
@@ -13,27 +13,30 @@ import { sexualOrientationValueToTitle } from '@/src/pages/create-game-page/ui/s
 import { characterBalanceValueToTitle } from '@/src/pages/create-game-page/ui/character-balance-value-to-title';
 import { GameOptionList } from '@/src/pages/create-game-page/ui/game-option-list/ui';
 import { renderBunkerSelectedText } from '@/src/pages/create-game-page/ui/render-bunker-selected-text';
-import { sheltersCategories } from '@/src/entities/shelter';
 import { ShelterCategoryList } from '@/src/shared/lib/types/shelter-category-list';
 import { shelterNameToShelter } from '@/src/entities/shelter/model/shelter-name-to-shelter';
-import { apocalypses, apocalypsesCategories } from '@/src/entities/apocalypse';
 import { ApocalypseCategories } from '@/src/shared/lib/types/apocalypse-categories';
 import { renderApocalypseSelectedText } from '@/src/pages/create-game-page/ui/render-apocalypse-selected-text';
 import { apocalypseNameToApocalypse } from '@/src/entities/apocalypse/model/apocalypse-name-to-apocalypse';
 import {
-  characteristicCards,
-  characteristicCardsList,
+  cardsStore,
 } from '@/src/entities/characteristic-card/model/characteristic-card';
-import { renderCharacteristicCardSelectedText } from '@/src/pages/create-game-page/ui/render-characteristic-card-selected-text';
+import {
+  renderCharacteristicCardSelectedText,
+} from '@/src/pages/create-game-page/ui/render-characteristic-card-selected-text';
 import { CharacteristicCardsList } from '@/src/shared/lib/types/characteristic-cards-list';
-import { characteristicCardNameToCard } from '@/src/entities/characteristic-card/model/characteristic-card-name-to-card';
-import { characteristicBalanceValueToTitle } from '@/src/pages/create-game-page/ui/characteristic-balance-value-to-title';
+import {
+  characteristicBalanceValueToTitle,
+} from '@/src/pages/create-game-page/ui/characteristic-balance-value-to-title';
 import { createGameStore } from '@/src/feature/create-game/model/create-game';
 import { gameStore } from '@/src/entities/game/model/game';
 import { OverlayModal } from '@/src/shared/ui/overlay-modal/ui';
-import { cardKitToCards } from '@/src/shared/lib/card-kit-to-cards';
 import { Image } from 'expo-image';
 import { adaptiveValue } from '@/src/shared/ui/adaptive-value/adaptive-value';
+import { apocalypsesStore } from '@/src/entities/apocalypse/model/apocalypses';
+import { sheltersStore } from '@/src/entities/shelter/model/shelters';
+import { databaseStore } from '@/src/shared/model/database-store';
+
 //TODO refactoring
 //TODO fix font issues
 
@@ -44,8 +47,14 @@ type PropsType = {
 export const CreateGamePage = observer(({ navigation }: PropsType) => {
   const [isErrorModalOpened, setIsErrorModalOpened] = useState<boolean>(false);
   const [errorDescription, setErrorDescription] = useState<string | null>(null);
-
+  const [isLoaded, setIsLoaded] = useState<boolean>();
   const { settings } = gameSettingsStore;
+
+  useEffect(() => {
+    setIsLoaded(true);
+  }, [settings]);
+
+  console.log(databaseStore.database.getAllSync(`SELECT name FROM sqlite_master WHERE type='table';`))
 
   return (
     <View style={s.createGamePageWrapper}>
@@ -97,101 +106,123 @@ export const CreateGamePage = observer(({ navigation }: PropsType) => {
                     gameSettingsStore.setSettings(options => (options.sexualOrientation = orientation))
                   }
                 />
-                <GameOptionRange
-                  title={'Уровень сложности'}
-                  defaultValue={settings.difficulty}
-                  description={'Параметр определяет насколько персонажи полезны и безопасны в среднем'}
-                  descriptionHeight={100}
-                  selectedTitle={difficultyValueToTitle(gameSettingsStore.settings.difficulty)}
-                  onValueChanged={difficulty =>
-                    gameSettingsStore.setSettings(options => (options.difficulty = difficulty))
-                  }
-                  min={gameSettingsStore.settingsLimits.difficulty.min}
-                  max={gameSettingsStore.settingsLimits.difficulty.max}
-                />
-                <GameOptionRange
-                  title={'Баланс Персонажей'}
-                  descriptionHeight={100}
-                  description={'Параметр определяет насколько различается полезность персонажей'}
-                  onValueChanged={characterBalance =>
-                    gameSettingsStore.setSettings(options => (options.balance = characterBalance))
-                  }
-                  selectedTitle={characterBalanceValueToTitle(settings.balance)}
-                  min={gameSettingsStore.settingsLimits.balance.min}
-                  max={gameSettingsStore.settingsLimits.balance.max}
-                  defaultValue={settings.balance}
-                />
-                <GameOptionRange
-                  title={'Баланс характеристик'}
-                  descriptionHeight={70}
-                  description={'Параметр определяет разброс характеристик'}
-                  onValueChanged={characteristicBalance =>
-                    gameSettingsStore.setSettings(settings => (settings.characteristicBalance = characteristicBalance))
-                  }
-                  selectedTitle={characteristicBalanceValueToTitle(settings.characteristicBalance)}
-                  min={gameSettingsStore.settingsLimits.characteristicBalance.min}
-                  max={gameSettingsStore.settingsLimits.characteristicBalance.max}
-                  defaultValue={settings.characteristicBalance}
-                />
 
-                <GameOptionList<ShelterCategoryList>
-                  title={'Список бункеров'}
-                  descriptionHeight={100}
-                  description={'Выберите из списка, с какими бункерами вы хотите играть'}
-                  renderSelectedText={renderBunkerSelectedText}
-                  items={sheltersCategories}
-                  uniqueKey={'name'}
-                  subKey={'children'}
-                  displayKey={'name'}
-                  selectText={'Выбрать бункер'}
-                  selectedByDefault={['Тора Бора']}
-                  searchPlaceholderText={'Искать Бункеры'}
-                  onValueChange={shelterNames =>
-                    gameSettingsStore.setSettings(options => {
-                      options.shelters = shelterNames.map(shelterName => shelterNameToShelter(shelterName));
-                    })
-                  }
-                />
-                <GameOptionList<ApocalypseCategories>
-                  title={'Список апокалипсисов'}
-                  descriptionHeight={100}
-                  description={'Выберите из списка, с какими апокалипсисами вы хотите играть'}
-                  renderSelectedText={renderApocalypseSelectedText}
-                  items={apocalypsesCategories}
-                  uniqueKey={'name'}
-                  displayKey={'name'}
-                  selectedByDefault={['Гигантские Змеи']}
-                  selectText={'Выбрать Апокалипсис'}
-                  searchPlaceholderText={'Искать Апокалипсисы'}
-                  onValueChange={apocalypsesNames =>
-                    gameSettingsStore.setSettings(options => {
-                      options.apocalypses = apocalypsesNames.map(apocalypseName =>
-                        apocalypseNameToApocalypse(apocalypseName),
-                      );
-                    })
-                  }
-                  subKey={'children'}
-                />
-                <GameOptionList<CharacteristicCardsList>
-                  title={'Список используемых карточек'}
-                  descriptionHeight={100}
-                  description={'Выберите из списка, с какими карточками вы хотите играть'}
-                  renderSelectedText={renderCharacteristicCardSelectedText}
-                  items={characteristicCardsList}
-                  uniqueKey={'name'}
-                  displayKey={'name'}
-                  selectText={'Выбрать карточки'}
-                  subKey={'children'}
-                  selectedByDefault={cardKitToCards(characteristicCards).map(card => card.name)}
-                  searchPlaceholderText={'Искать карточки'}
-                  onValueChange={characteristicCardsNames =>
-                    gameSettingsStore.setSettings(options => {
-                      options.cardsKit = characteristicCardsNames.map(cardName =>
-                        characteristicCardNameToCard(cardName),
-                      );
-                    })
-                  }
-                />
+                {isLoaded && <>
+
+                  <GameOptionRange
+                    title={'Уровень сложности'}
+                    defaultValue={settings.difficulty}
+                    description={'Параметр определяет насколько персонажи полезны и безопасны в среднем'}
+                    descriptionHeight={100}
+                    selectedTitle={difficultyValueToTitle(gameSettingsStore.settings.difficulty)}
+                    onValueChanged={difficulty =>
+                      gameSettingsStore.setSettings(options => (options.difficulty = difficulty))
+                    }
+                    min={gameSettingsStore.settingsLimits.difficulty.min}
+                    max={gameSettingsStore.settingsLimits.difficulty.max}
+                  />
+
+                  <GameOptionRange
+                    title={'Баланс Персонажей'}
+                    descriptionHeight={100}
+                    description={'Параметр определяет насколько различается полезность персонажей'}
+                    onValueChanged={characterBalance =>
+                      gameSettingsStore.setSettings(options => (options.balance = characterBalance))
+                    }
+                    selectedTitle={characterBalanceValueToTitle(settings.balance)}
+                    min={gameSettingsStore.settingsLimits.balance.min}
+                    max={gameSettingsStore.settingsLimits.balance.max}
+                    defaultValue={settings.balance}
+                  />
+                  <GameOptionRange
+                    title={'Баланс характеристик'}
+                    descriptionHeight={70}
+                    description={'Параметр определяет разброс характеристик'}
+                    onValueChanged={characteristicBalance =>
+                      gameSettingsStore.setSettings(settings => (settings.characteristicBalance = characteristicBalance))
+                    }
+                    selectedTitle={characteristicBalanceValueToTitle(settings.characteristicBalance)}
+                    min={gameSettingsStore.settingsLimits.characteristicBalance.min}
+                    max={gameSettingsStore.settingsLimits.characteristicBalance.max}
+                    defaultValue={settings.characteristicBalance}
+                  />
+
+                  <GameOptionList<ShelterCategoryList>
+                    title={'Список бункеров'}
+                    descriptionHeight={100}
+                    description={'Выберите из списка, с какими бункерами вы хотите играть'}
+                    renderSelectedText={renderBunkerSelectedText}
+                    items={[
+                      {
+                        name: 'Бункеры Стандартного Издания',
+                        children: sheltersStore.getSheltersByEdition('standart'),
+                      },
+                      {
+                        name: 'Бункеры Издания "Боров"',
+                        children: sheltersStore.getSheltersByEdition('borov'),
+                      },
+                    ]}
+                    uniqueKey={'name'}
+                    subKey={'children'}
+                    displayKey={'name'}
+                    selectText={'Выбрать бункер'}
+                    selectedByDefault={['Тора Бора']}
+                    searchPlaceholderText={'Искать Бункеры'}
+                    onValueChange={shelterNames =>
+                      gameSettingsStore.setSettings(options => {
+                        options.shelters = shelterNames.map(shelterName => shelterNameToShelter(shelterName));
+                      })
+                    }
+                  />
+                  <GameOptionList<ApocalypseCategories>
+                    title={'Список апокалипсисов'}
+                    descriptionHeight={100}
+                    description={'Выберите из списка, с какими апокалипсисами вы хотите играть'}
+                    renderSelectedText={renderApocalypseSelectedText}
+                    items={[
+                      {
+                        name: 'Апокалипсисы Стандартного Издания',
+                        children: apocalypsesStore.getApocalypsesByEdition('standart'),
+                      },
+                      {
+                        name: 'Апокалипсисы Издания "Боров"',
+                        children: apocalypsesStore.getApocalypsesByEdition('borov'),
+                      },
+                    ]}
+                    uniqueKey={'name'}
+                    displayKey={'name'}
+                    selectedByDefault={['Гигантские Змеи']}
+                    selectText={'Выбрать Апокалипсис'}
+                    searchPlaceholderText={'Искать Апокалипсисы'}
+                    onValueChange={apocalypsesNames =>
+                      gameSettingsStore.setSettings(options => {
+                        options.apocalypses = apocalypsesNames.map(apocalypseNameToApocalypse);
+                      })
+                    }
+                    subKey={'children'}
+                  />
+                  <GameOptionList<CharacteristicCardsList>
+                    title={'Список используемых карточек'}
+                    descriptionHeight={100}
+                    description={'Выберите из списка, с какими карточками вы хотите играть'}
+                    renderSelectedText={renderCharacteristicCardSelectedText}
+                    items={[
+                      { name: 'Карточки Стандартного Издания', children: cardsStore.getCardsByEdition('standart') },
+                      { name: 'Карточки Издания "Боров"', children: cardsStore.getCardsByEdition('borov') },
+                    ]}
+                    uniqueKey={'name'}
+                    displayKey={'name'}
+                    selectText={'Выбрать карточки'}
+                    subKey={'children'}
+                    selectedByDefault={cardsStore.getAllCards().map(card => card.name)}
+                    searchPlaceholderText={'Искать карточки'}
+                    onValueChange={characteristicCardsNames =>
+                      gameSettingsStore.setSettings(options => options.cardsKit = characteristicCardsNames.map(cardsStore.getCardByName))
+
+                    }
+                  />
+                </>}
+
               </ScrollView>
             </SafeAreaView>
           </View>