+++ src/shared/lib/types/shelter-encoded.ts
@@ -1,0 +1,6 @@
+import { Shelter } from '@/src/shared/lib/types/shelter';
+
+export type ShelterEncoded ={
+  rooms:string,
+  resources:string
+} &  Shelter